requirements.txt
faiss-cpu
sentence-transformers
langchain>=0.1.0
langchain-community
python-dotenv
numpy>=1.25.0
torch>=2.1.0
google-genai
fastapi
uvicorn